QA for rooms and photos
